+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+- Improve panic message on stable.
+
 ## [0.1.0] - 2023-03-22
 
 Initial release